refactor : 함수 클래스 이동 (#731)
